+++ sdk/helper/ptr/ptr.go
@@ -8,6 +8,10 @@ func IntToPtr(i int) *int {
 	return &i
 }
 
+func Int32ToPtr(i int32) *int32 {
+	return &i
+}
+
 func Int64ToPtr(i int64) *int64 {
 	return &i
 }